The current domain XML used for the KVM driver does not specify a CPU mode, which makes things very slow and usually lead to a constant >25% CPU usage for people using your driver with Kubernetes and Minikube. This solves the problem as it simply tells KVM to use everything the host supports and emulate whatever else is needed.
See http://wiki.qemu.org/Features/CPUModels#-cpu_host_and_feature_probing for more info. CPU host mode is usually safe and recommended for scenarios like the ones this driver is used for. The only limitation (which I suspect is not a big deal here) is that if KVM migrations between hosts of different CPUs models are too frequent.