Modify Makefile and Dockerfiles to speed up builds
Build a base docker image for each platform with the requisite go
executable, libvirt packages, and docker-machine source. When making
releases, this base image will be used to build a second temporary
image in which the source directory is mounted. This image is then used
to launch a container in which the executable is built, after which
both container and build image are removed.

This speeds up compilation of the executables for each platform and
allows for testing changes on all platforms more quickly